+++ bas_metadata_library/standards/iso_19115_1/__init__.py
@@ -165,3 +165,7 @@ class MetadataRecord(_MetadataRecord):
 
     def make_element(self) -> Element:
         return self.metadata_record.make_element()
+
+    # noinspection PyMethodOverriding
+    def validate(self) -> None:
+        super().validate(xsd_path=Path("gmd/gmd.xsd"))